+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] -
 
+### Added
+
+- Parameter to change label to given image.
+
+### Changed
+
+- Label and overview series can now be None.
+
 ## [0.6.0] - 2023-01-24
 
 ### Changed